Android: Merging AndroidInputThreePlus into AndroidInput.
